Conflicts:
	Robot Flier.depend
	build/Robot Flier.exe
	main.cpp
	output/main.o

+++ globals.cpp
@@ -4,8 +4,8 @@
 const int robotWidth = 70;
 const int robotHeight = 70;
 
-int robot_x;
-int robot_y;
+float robot_x;
+float robot_y;
 
 int health;
 unsigned int score;
